requirements.txt
pydantic
pillow
transformers
requests
paddleocr
uvicorn
gunicorn